增加了删除时的弹窗提示, 并且将登陆注册页面做了居中优化

+++ src/main/webapp/Project_declaration_list.jsp
@@ -253,13 +253,35 @@
           <td>${research.reporting_of_results}</td>
           <td>
             <a href="${pageContext.request.contextPath}/Scientific_research_project/declare_find_by_id?id=${research.id}" class="button">编辑</a>
-            <a href="${pageContext.request.contextPath}/Scientific_research_project/declare_delete?id=${research.id}" class="button">删除</a>
+            <a href="${pageContext.request.contextPath}/Scientific_research_project/declare_delete?id=${research.id}" onclick="confirmDelete(event)" class="button">删除</a>
 <%--            <a class="button">成果汇报</a>--%>
           </td>
         </tr>
       </c:forEach>
 
     </table>
+
+    <script type="text/javascript">
+      function confirmDelete(event) {
+        // 阻止默认的链接点击行为
+        event.preventDefault();
+
+        // 弹出确认对话框
+        var result = confirm("确定要删除吗？");
+
+        // 根据用户的选择执行相应操作
+        if (result) {
+          // 用户点击了确认，执行删除操作
+          var deleteUrl = event.target.href;
+          window.location.href = deleteUrl;
+        } else {
+          // 用户点击了取消，不执行任何操作
+          return false;
+        }
+      }
+    </script>
+
+
   </section>
 </main>
 </body>
